+++ src/Freya.Machine.Extensions.Http/Operations.fs
@@ -22,6 +22,7 @@
 module internal Freya.Machine.Extensions.Http.Operations
 
 open System
+open Arachne.Http
 open Freya.Core
 open Freya.Core.Operators
 open Freya.Machine