+++ tests/Freya.Machine.Tests/Prelude.fs
@@ -23,13 +23,14 @@ module Freya.TodoBackend.Prelude
 
 open System.IO
 open System.Text
+open Arachne.Http
+open Arachne.Language
 open Chiron
 open Freya.Core
 open Freya.Core.Operators
 open Freya.Machine
 open Freya.Machine.Extensions.Http
 open Freya.Types.Http
-open Freya.Types.Language
 
 (* Utility