+++ tests/Freya.Router.Tests/Prelude.fs
@@ -4,6 +4,7 @@ module internal Freya.Router.Tests.Prelude
 open System.Collections.Generic
 open Aether
 open Aether.Operators
+open Arachne.Http
 open Freya.Core
 open Freya.Core.Operators
 open Freya.Router